% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsSeeds.R
\name{confidenceBands}
\alias{confidenceBands}
\title{Get the estimated confidence bands for the bayesian method}
\usage{
confidenceBands(resultsSeeds, slot, ind)
}
\arguments{
\item{resultsSeeds}{A object of the class 'resultsSeeds', which is returned from the algorithms.}

\item{slot}{Arguement of type character.}

\item{ind}{A numeric indicating the index of a 'resultsSeeds'-Object in a list. If not set the last listed object will be used.}
}
\description{
Get the estimated confidence bands for the bayesian method
}
