% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootEvents.R
\name{createCLangRoot}
\alias{createCLangRoot}
\title{A function for creating a root function that can be used with the ode solver lsoda for triggering events - C Version}
\usage{
createCLangRoot(rootStates)
}
\arguments{
\item{rootStates}{Vector of the same length as the state, ones and zeros indicated which states can start an event}
}
\description{
A function for creating a root function that can be used with the ode solver lsoda for triggering events - C Version
}
