% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootEvents.R
\name{createEvent}
\alias{createEvent}
\title{Create an event function that sets states that are zero to a smalll value}
\usage{
createEvent(tollerance, value)
}
\arguments{
\item{tollerance}{tollerance under which a state is set to a given value 'value'}

\item{value}{value the states should be set to, very small values result in a much longer computation time}
}
\description{
Create an event function that sets states that are zero to a smalll value
}
