% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFunctions.R
\name{createFunctions}
\alias{createFunctions}
\title{creates files for costate and argumented state}
\usage{
createFunctions(odeEq)
}
\arguments{
\item{odeEq}{a object of the odeEquation class}
}
\description{
The function handles the formating of the ODE
}
