% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIBBS_update.R
\name{GIBBS_update}
\alias{GIBBS_update}
\title{calculates the Gibbs Update}
\usage{
GIBBS_update(D, EPS_inner, R, ROH, SGIMA_0, n, SIGMA, LAMBDA2, LAMBDA1, TAU)
}
\description{
calculates the Gibbs Update
}
