% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSBML.R
\name{importSBML}
\alias{importSBML}
\title{Import SBML Models using the Bioconductor package 'rsbml'}
\usage{
importSBML(filename, times, y)
}
\description{
A simple function for importing sbml models from a extensive markup language file.
 
 @param  filename name of the import file. Should lay in the working directory.
 @param  times     timestep at which the function should be evaluated
 @param  y measurements of the model
}
\details{
@export importSBML
}
