% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jakstatInput}
\alias{jakstatInput}
\title{Input of the Jak2/Stat5 signaling pathway}
\format{An object of class \code{data.frame} with 16 rows and 2 columns.}
\source{
\url{http://msb.embopress.org/content/7/1/516}
}
\usage{
jakstatInput
}
\description{
A dataframe that containing the inputs of the Jak2/Stat5 pathway that is
used to stimulate the system. The data is taken
from the works of Timmer et al. (DOI 10.1038/msb.2011.50).
}
\details{
A dataframe with 16 rows and 2 columns
\describe{
   \item{t}{t in minutes}
   \item{u}{activation of Epo-receptor}
}
}
\keyword{datasets}
