% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jakstatMeasurement}
\alias{jakstatMeasurement}
\title{Data from the JakStat signaling}
\format{An object of class \code{data.frame} with 16 rows and 5 columns.}
\source{
\url{http://msb.embopress.org/content/7/1/516}
}
\usage{
jakstatMeasurement
}
\description{
This is the data from the Jak2/Stat5 signaling pathway. The data is taken
from the works of Timmer et al. (DOI 10.1038/msb.2011.50).
}
\details{
A data frame with 16 rows and 5 columns
\describe{
   \item{t}{t in minutes}
   \item{y1}{phosphorylated STAT5 in cytoplasma}
   \item{y1sd}{standard deviation of the measurement of y1 at time t}
   \item{y2}{total STAT5 in cytoplama}
   \item{y2sd}{standard deviation of the measurement of y2 at time t}
}
}
\keyword{datasets}
