% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood_Function.R
\name{LOGLIKELIHOOD_func}
\alias{LOGLIKELIHOOD_func}
\title{calculates the log likelihood log[L(G|x)P(G)]}
\usage{
LOGLIKELIHOOD_func(pars, Step, OBSERVATIONS, x_0, parameters, EPS_inner, INPUT,
  D, GIBBS_PAR, k, MU_JUMP, SIGMA_JUMP, eps_new, objectivfunc)
}
\description{
calculates the log likelihood log[L(G|x)P(G)]
}
