% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_component.R
\name{MCMC_component}
\alias{MCMC_component}
\title{mcmc sampler}
\usage{
MCMC_component(LOGLIKELIHOOD_func, STEP_SIZE, STEP_SIZE_INNER, EPSILON,
  JUMP_SCALE, STEP, OBSERVATIONS, Y0, INPUTDATA, PARAMETER, EPSILON_ACT, SIGMA,
  DIAG, GIBBS_par, N, BURNIN, objective)
}
\description{
mcmc sampler
}
