% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bden_uvb.R
\docType{data}
\name{Model}
\alias{Model}
\title{Test dataset  for demonstrating the bden algorithm.}
\format{
An object of class \code{odeModel} of length 1.
}
\usage{
data(bden_uvb)
}
\description{
Dataset is identical with the example for the bden algorithm from the vignette. It contains an object of
odeModel that describes the uvb network.
}
\keyword{datasets}
