% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelClass.R
\docType{methods}
\name{setModelEquation,odeModel-method}
\alias{setModelEquation,odeModel-method}
\alias{setParms,odeModel-method}
\alias{setInput,odeModel-method}
\alias{setMeasFunc,odeModel-method}
\alias{setY,odeModel-method}
\alias{setMeas,odeModel-method}
\alias{setSd,odeModel-method}
\alias{nominalSol,odeModel,missing-method}
\title{Set the model equation}
\usage{
\S4method{setModelEquation}{odeModel}(theObject, func)

\S4method{setParms}{odeModel}(theObject, parms)

\S4method{setInput}{odeModel}(theObject, input)

\S4method{setMeasFunc}{odeModel}(theObject, measFunc)

\S4method{setY}{odeModel}(theObject, y)

\S4method{setMeas}{odeModel}(theObject, meas)

\S4method{setSd}{odeModel}(theObject, sd)

\S4method{nominalSol}{odeModel,missing}(odeModel, logTrans)
}
\arguments{
\item{theObject}{an object of the class modelClass}

\item{func}{function describing the ode equation of the model}

\item{parms}{a vector containing the parmeters of the model}

\item{measFunc}{measurement function of the model. Has to be a R functions.}

\item{y}{vector with the initial values}

\item{meas}{measurements of the model, a matrix with measurements of the model
and the corresponding time values}

\item{sd}{a matrix with the standard deviations of the measurements}

\item{odeModel}{a object of the class ode model describing the experiment}

\item{logTrans}{a vector indicating which of the state vector componenets should be log transformed}

\item{theObject}{an object of the class modelClass}

\item{theObject}{an object of the class modelClass}

\item{func}{function describing the ode equation of the model}

\item{theObject}{an object of the class modelClass}

\item{theObject}{an object of the class modelClass}

\item{theObject}{an object of the class modelClass}

\item{theObject}{an object of the class modelClass}
}
\description{
Set the model equation of the system. Has to be a function that can be used with the deSolve package
}
