% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelClass.R
\name{nominalSol}
\alias{nominalSol}
\title{Calculate the nominal solution of the model}
\usage{
nominalSol(odeModel, logTrans)
}
\arguments{
\item{odeModel}{a object of the class ode model describing the experiment}

\item{logTrans}{a vector indicating which of the state vector componenets should be log transformed}
}
\description{
Calculate the nominal solution of the model
}
