% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classOdeEquation.R
\docType{class}
\name{odeEquations-class}
\alias{odeEquations-class}
\alias{odeEq}
\title{An S4 class used to handle formating ODE-Equation and calculate the needed functions for the SEEDS-algorithm}
\description{
An S4 class used to handle formating ODE-Equation and calculate the needed functions for the SEEDS-algorithm
}
\section{Slots}{

\describe{
\item{\code{modelStr}}{a vector of strings describing the ODE}

\item{\code{measureStr}}{a vector of strings describing the}

\item{\code{origEq}}{a vector of strings containing the original model function}

\item{\code{measureFunction}}{a vector of strings containing the original measurement function}

\item{\code{costateEq}}{a vector of strings describing the costate equation}

\item{\code{JhT}}{a matrix of strings describing the jacobian matrix of the measurement function}

\item{\code{jacobian}}{a matrix of strings representing the jacobian matrix model equations}

\item{\code{costFunction}}{a string containing the cost function}

\item{\code{hamiltonian}}{a string representing the Hamilton function of the model}
}}

