% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsSeeds.R
\name{outputEstimates}
\alias{outputEstimates}
\alias{outputEstimates,list,numeric-method}
\alias{outputEstimates,list,missing-method}
\alias{outputEstimates,resultsSeeds,missing-method}
\title{Get the estimated outputs}
\usage{
outputEstimates(resultsSeeds, ind)

\S4method{outputEstimates}{list,numeric}(resultsSeeds, ind)

\S4method{outputEstimates}{list,missing}(resultsSeeds, ind)

\S4method{outputEstimates}{resultsSeeds,missing}(resultsSeeds, ind)
}
\arguments{
\item{resultsSeeds}{A object of the class 'resultsSeeds', which is returned from the algorithms.}

\item{ind}{A numeric indicating the index of a 'resultsSeeds'-Object in a list. If not set the last listed object will be used.}
}
\value{
Dafaframe with estimated measurements.
}
\description{
Get the estimated outputs
}
\examples{

data(uvb_res)

outputEstimates(res[[2]])

}
