% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds_circle.R
\name{plot_circle}
\alias{plot_circle}
\title{plot circle}
\usage{
plot_circle(r = 1, ...)
}
\arguments{
\item{r}{circle's radius you want to plot}

\item{...}{ggplot's option}
}
\description{
plot circle
}
\examples{
# ggplot circle
plot_circle()
}
