% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds_cycl.R
\name{plot_cycl}
\alias{plot_cycl}
\title{plot cycloid}
\usage{
plot_cycl(radius_m = 1, ...)
}
\arguments{
\item{radius_m}{movable circle (Small circle)'s radius}

\item{...}{ggplot's option}
}
\description{
plot cycloid
}
\examples{
# ggplot cycloid
plot_cycl(3)
}
