% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds_descartes.R
\name{plot_descartes}
\alias{plot_descartes}
\title{plot descartes curve}
\usage{
plot_descartes(a = 1, ...)
}
\arguments{
\item{a}{coefficient}

\item{...}{ggplot's option}
}
\description{
plot descartes curve
}
\examples{
# ggplot hypocycloid
plot_descartes(1)
}
