% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds_ellipse.R
\name{plot_ellipse}
\alias{plot_ellipse}
\title{plot ellipse}
\usage{
plot_ellipse(a = 1, b = 1, ...)
}
\arguments{
\item{a}{sin's coefficient}

\item{b}{cos's coefficient}

\item{...}{ggplot's option}
}
\description{
plot ellipse
}
\examples{
# ggplot ellipse
plot_ellipse(2, 3)
}
