% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds_epicycl.R
\name{plot_epicycl}
\alias{plot_epicycl}
\title{plot epicycloid}
\usage{
plot_epicycl(radius_c = 1, radius_m = 1, ...)
}
\arguments{
\item{radius_c}{constant circle (Large circle)'s radius}

\item{radius_m}{movable circle (Small circle)'s radius}

\item{...}{ggplot's option}
}
\description{
plot epicycloid
}
\examples{
# ggplot hypocycloid
plot_epicycl(1, 3)
}
