% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds_hypocycl.R
\name{plot_hypocycl}
\alias{plot_hypocycl}
\title{plot hypocycloid}
\usage{
plot_hypocycl(radius_c = 1, radius_m = 1, ...)
}
\arguments{
\item{radius_c}{constant circle (Large circle)'s radius}

\item{radius_m}{movable circle (Small circle)'s radius}

\item{...}{ggplot's option}
}
\description{
plot hypocycloid
}
\examples{
# ggplot hypocycloid
plot_hypocycl(1, 3)
}
