% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds_log_spiral.R
\name{plot_log_spiral}
\alias{plot_log_spiral}
\title{plot log_spiral}
\usage{
plot_log_spiral(a = 1, b = 1, ...)
}
\arguments{
\item{a}{coefficient}

\item{b}{coefficient}

\item{...}{ggplot's option}
}
\description{
plot log_spiral
}
\examples{
# ggplot log_spiral
plot_log_spiral(2, 3)
}
