% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds_param.R
\name{plot_param}
\alias{plot_param}
\title{plot parameter representation}
\usage{
plot_param(a = 1, b = 1, nsin = 1, ncos = 0, k = 1, ...)
}
\arguments{
\item{a}{coefficient of sin}

\item{b}{coefficient of cos}

\item{nsin}{coefficient of sin's radian}

\item{ncos}{coefficient of cos's radian}

\item{k}{constant number}

\item{...}{ggplot's option}
}
\description{
plot parameter representation
}
\examples{
# ggplot parameter representation
plot_param(2, 3, 4, 5, 2)
}
