% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds_pascal.R
\name{plot_pascal}
\alias{plot_pascal}
\title{plot pascal curve}
\usage{
plot_pascal(a = 1, l = 1, ...)
}
\arguments{
\item{a}{coefficient}

\item{l}{constant}

\item{...}{ggplot's option}
}
\description{
plot pascal curve
}
\examples{
# ggplot hypocycloid
plot_pascal(1)
}
