% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds_rose.R
\name{plot_rose}
\alias{plot_rose}
\title{plot rose curve}
\usage{
plot_rose(a = 1, nsin = 1, k = 1, ...)
}
\arguments{
\item{a}{coefficient of sin}

\item{nsin}{coefficient of sin's radian}

\item{k}{constant number}

\item{...}{ggplot's option}
}
\description{
plot rose curve
}
\examples{
# ggplot hypocycloid
plot_rose(1, 2, 1)
}
