% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsSeeds.R
\docType{methods}
\name{plotAnno}
\alias{plotAnno}
\alias{plotAnno,resultsSeeds-method}
\alias{plotAnno,list-method}
\title{Create annotated plot}
\usage{
plotAnno(x, stateAnno, measAnno)

\S4method{plotAnno}{resultsSeeds}(x, stateAnno, measAnno)

\S4method{plotAnno}{list}(x, stateAnno, measAnno)
}
\arguments{
\item{x}{an object of type resultsSeeds which contains the results of the algorithms}

\item{stateAnno}{a character vector describing the names of the states}

\item{measAnno}{a character vector describing the names of the measurements}
}
\description{
Create a annotated plot with given state and measurement names. The plots are
equal to the output of the normal plot function.
}
