% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsSeeds.R
\docType{methods}
\name{plotHiddenInputs}
\alias{plotHiddenInputs}
\alias{plotHiddenInputs,resultsSeeds,missing-method}
\title{Plot the estimated states}
\usage{
plotHiddenInputs(x, y)

\S4method{plotHiddenInputs}{resultsSeeds,missing}(x, y)
}
\arguments{
\item{resultsSeeds}{}
}
\description{
Plot the estimated states
}
