% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsSeeds.R
\name{plot,resultsSeeds,missing-method}
\alias{plot,resultsSeeds,missing-method}
\title{Plot method for the S4 class resultsSeeds}
\usage{
\S4method{plot}{resultsSeeds,missing}(x, y)
}
\arguments{
\item{x}{an object of type resultsSeeds or a list of these objects. If a list
is given the last entry will be plotted.}

\item{y}{...}
}
\value{
A list of plots showing the results of the algorithm
}
\description{
A standardized plot function to display the results of the algorithms. Both
algorithms should result in objects of the class resultsSeeds. The results can
be plotted using the \code{\link{plot}}-function.
}
\examples{

data(uvb_res)

plot(res[[2]])

}
