% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsSeeds.R
\docType{methods}
\name{plotStates}
\alias{plotStates}
\alias{plotStates,resultsSeeds,missing-method}
\alias{plotMeas}
\alias{plotMeas,resultsSeeds,missing-method}
\title{Plot the estimated states}
\usage{
plotStates(x, y)

\S4method{plotStates}{resultsSeeds,missing}(x, y)

plotMeas(x, y)

\S4method{plotMeas}{resultsSeeds,missing}(x, y)
}
\arguments{
\item{resultsSeeds}{}

\item{resultsSeeds}{}
}
\description{
Plot the estimated states

Plot the estimated states
}
