% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvb-res-data.R
\docType{data}
\name{res}
\alias{res}
\title{Results from the uvb dataset for examples}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(uvb_res)
}
\description{
Data from running the estimation of hidden inputs from the UVB-G Protein demo. This data is used for demonstration the different functions of the package
}
\keyword{datasets}
