% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedyApproach.R, R/seeds.R
\docType{package}
\name{seeds-package}
\alias{seeds}
\alias{seeds-package}
\alias{seeds}
\alias{seeds-package}
\title{Algorithms for calculating the hidden inputs of a system}
\description{
SEEDS combines two algorithms to estimate the hidden inputs of a nominal system
with given measurements.

A package combining two different algorithms for calculating hidden
inputs. These hidden inputs can be interpreted as a controll that tries to minimize the
discrepancies between a given model and a given measurements. The method is
also called the 'dynamic elastic net'. The related work can be found in doi:10.1038/srep20772
}
\details{
This package combines two algorithms for calculating the hidden inputs. A more
simple approach using optimal control based on the maximum principal using a
conjugate gradient algorithm and a MCMC-method.

Details
}
\author{
\strong{Maintainer}: Tobias Newmiwaka \email{tobias.newmiwaka@gmail.com}


\strong{Maintainer}: Tobias Newmiwaka \email{tobias.newmiwaka@gmail.com}

}
