% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds_circle.R
\name{seeds_circle_df}
\alias{seeds_circle_df}
\alias{seeds_circle_tibble}
\title{for plot circle, generate data.frame or tibble}
\usage{
seeds_circle_df(r = 1)

seeds_circle_tibble(r = 1)
}
\arguments{
\item{r}{circle's radius you want to plot}
}
\description{
for plot circle, generate data.frame or tibble
}
\examples{
# return data.frame(row number is 100)
seeds_circle_df()

# return tibble(row number is 100)
seeds_circle_tibble()
}
