% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds_descartes.R
\name{seeds_descartes_df}
\alias{seeds_descartes_df}
\alias{seeds_descartes_tibble}
\title{for plot descartes curve, generate data.frame of tibble}
\usage{
seeds_descartes_df(a = 1)

seeds_descartes_tibble(a = 1)
}
\arguments{
\item{a}{coefficient of sin}
}
\description{
for plot descartes curve, generate data.frame of tibble
}
\examples{
# return data.frame(row number is 1000)
seeds_descartes_df(1)

# return tibble(row number is 1000)
seeds_descartes_tibble(1)
}
