% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds_ellipse.R
\name{seeds_ellipse_df}
\alias{seeds_ellipse_df}
\alias{seeds_ellipse_tibble}
\title{for plot ellipse, generate data.frame or tibble}
\usage{
seeds_ellipse_df(a = 1, b = 1)

seeds_ellipse_tibble(a = 1, b = 1)
}
\arguments{
\item{a}{sin's coefficient}

\item{b}{cos's coefficient}
}
\description{
for plot ellipse, generate data.frame or tibble
}
\examples{
# return data.frame(row number is 100)
seeds_ellipse_df(2, 3)

# return tibble(row number is 100)
seeds_ellipse_tibble(2, 3)
}
