% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds_log_spiral.R
\name{seeds_log_spiral_df}
\alias{seeds_log_spiral_df}
\alias{seeds_log_spiral_tibble}
\title{for plot logarithmic spiral, generate data.frame or tibble}
\usage{
seeds_log_spiral_df(a = 1, b = 1)

seeds_log_spiral_tibble(a = 1, b = 1)
}
\arguments{
\item{a}{coefficient}

\item{b}{coefficient}
}
\description{
for plot logarithmic spiral, generate data.frame or tibble
}
\examples{
# return data.frame(row number is 100)
seeds_log_spiral_df(2, 3)

# return tibble(row number is 100)
seeds_log_spiral_tibble(2, 3)
}
