% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds_param.R
\name{seeds_param_df}
\alias{seeds_param_df}
\alias{seeds_param_tibble}
\title{for plot parameter representation, generate data.frame of tibble}
\usage{
seeds_param_df(a = 1, b = 1, nsin = 1, ncos = 0, k = 1)

seeds_param_tibble(a = 1, b = 1, nsin = 1, ncos = 0, k = 1)
}
\arguments{
\item{a}{coefficient of sin}

\item{b}{coefficient of cos}

\item{nsin}{coefficient of sin's radian}

\item{ncos}{coefficient of cos's radian}

\item{k}{constant number}
}
\description{
for plot parameter representation, generate data.frame of tibble
}
\examples{
# return data.frame(row number is 1000)
seeds_param_df(2, 3, 4, 5, 2)

# return tibble(row number is 1000)
seeds_param_tibble(2, 3, 4, 5, 2)
}
