% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds_pascal.R
\name{seeds_pascal_df}
\alias{seeds_pascal_df}
\alias{seeds_pascal_tibble}
\title{for plot limacon of pascal, generate data.frame of tibble}
\usage{
seeds_pascal_df(a = 1, l = 1)

seeds_pascal_tibble(a = 1, l = 1)
}
\arguments{
\item{a}{coefficient of sin}

\item{l}{constant}
}
\description{
for plot limacon of pascal, generate data.frame of tibble
}
\examples{
# return data.frame(row number is 1000)
seeds_pascal_df(1, 2)

# return tibble(row number is 1000)
seeds_pascal_tibble(1)
}
