% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds_rose.R
\name{seeds_rose_df}
\alias{seeds_rose_df}
\alias{seeds_rose_tibble}
\title{for plot rose curve, generate data.frame of tibble}
\usage{
seeds_rose_df(a = 1, nsin = 1, k = 1)

seeds_rose_tibble(a = 1, nsin = 1, k = 1)
}
\arguments{
\item{a}{coefficient of sin}

\item{nsin}{coefficient of sin's radian}

\item{k}{constant number}
}
\description{
for plot rose curve, generate data.frame of tibble
}
\examples{
# return data.frame(row number is 1000)
seeds_rose_df(1, 2, 1)

# return tibble(row number is 1000)
seeds_rose_tibble(1, 2, 1)
}
