% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelClass.R
\name{setInitState}
\alias{setInitState}
\alias{setInitState,odeModel-method}
\title{Set the vector with the initial (state) values}
\usage{
setInitState(odeModel, y)

\S4method{setInitState}{odeModel}(odeModel, y)
}
\arguments{
\item{odeModel}{an object of the class odeModel}

\item{y}{vector with the initial values}
}
\value{
an object of odeModel
}
\description{
Set the vector with the initial (state) values
}
\examples{

data("uvbModel")

x0 = c(0.2,10,2,0,0,20,0,0,0,4.2,0.25,20,0)

newModel <- setInitState(uvbModel, y = x0)

}
