% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelClass.R
\name{setInput}
\alias{setInput}
\title{Set the inputs of the model.}
\usage{
setInput(theObject, input)
}
\arguments{
\item{theObject}{an object of the class modelClass}

\item{func}{function describing the ode equation of the model}
}
\description{
Set the inputs of the model.
}
