% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelClass.R
\docType{methods}
\name{setMeas}
\alias{setMeas}
\alias{setMeas,odeModel-method}
\title{set measurements of the model}
\usage{
setMeas(theObject, meas)

\S4method{setMeas}{odeModel}(theObject, meas)
}
\arguments{
\item{theObject}{an object of the class odeModel}

\item{meas}{measurements of the model, a matrix with measurements of the model
and the corresponding time values}
}
\description{
set measurements of the model
}
