% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelClass.R
\docType{methods}
\name{setMeasFunc}
\alias{setMeasFunc}
\alias{setMeasFunc,odeModel,function,missing-method}
\alias{setMeasFunc,odeModel,function,logical-method}
\title{Set the measurement equation for the model}
\usage{
setMeasFunc(theObject, measFunc, costum)

\S4method{setMeasFunc}{odeModel,`function`,missing}(theObject, measFunc,
  costum)

\S4method{setMeasFunc}{odeModel,`function`,logical}(theObject, measFunc,
  costum)
}
\arguments{
\item{theObject}{an object of the class odeModel}

\item{measFunc}{measurement function of the model. Has to be a R functions.}

\item{costum}{costum indexing for the measurement function}
}
\description{
Set the measurement equation for the model
}
