% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelClass.R
\docType{methods}
\name{setModelEquation}
\alias{setModelEquation}
\alias{setModelEquation,odeModel-method}
\title{Set the model equation}
\usage{
setModelEquation(theObject, func)

\S4method{setModelEquation}{odeModel}(theObject, func)
}
\arguments{
\item{theObject}{an object of the class odeModel}

\item{func}{function describing the ode equation of the model}
}
\description{
Set the model equation of the system. Has to be a function that can be used with the deSolve package
}
