% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelClass.R
\docType{methods}
\name{setParms}
\alias{setParms}
\alias{setParms,odeModel,numeric-method}
\title{Set the model parameters}
\usage{
setParms(theObject, parms)

\S4method{setParms}{odeModel,numeric}(theObject, parms)
}
\arguments{
\item{theObject}{an object of the class odeModel}

\item{parms}{a vector containing the parameters of the model}
}
\description{
a method to set the model parameters of an odeModel object.
}
