% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelClass.R
\name{setSd}
\alias{setSd}
\title{Set the standard deviation of the measurements}
\usage{
setSd(theObject, sd)
}
\arguments{
\item{theObject}{an object of the class odeModel}

\item{sd}{a matrix with the standard deviations of the measurements}
}
\description{
Set the standard deviation of the measurements
}
