% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolicDiff.R
\name{symbolicDiff}
\alias{symbolicDiff}
\title{Symbolic differentiation to create the adjoint equations}
\usage{
symbolicDiff(odeObj)
}
\arguments{
\item{odeObj}{a object of the S4 class odeEquation}
}
\value{
a list of the costate equation and a jacobian matrix
}
\description{
The function calculates the needed equations for the gradient descent algorithm.
The function needs a object of the S4 OdeEquation class. This class contains
all the needed equations.
}
\details{
The calculation is based on the Derive function of the package Deriv. The needed
object can be created with the S4 class OdeEquations.
}
