% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvb-model-obj.R
\docType{data}
\name{uvbModel}
\alias{uvbModel}
\title{An object of the odeModel Class}
\format{
An object of class \code{odeModel} of length 1.
}
\usage{
data(uvbModel)
}
\description{
Object is used for demonstrating the functions of the odeModel Class. It is used in the demos for the uvb signaling pathway.
}
\keyword{datasets}
