% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsSeeds.R
\docType{methods}
\name{confidenceBands}
\alias{confidenceBands}
\alias{confidenceBands,list,character,numeric-method}
\alias{confidenceBands,list,character,missing-method}
\alias{confidenceBands,resultsSeeds,character,missing-method}
\title{Get the estimated confidence bands for the bayesian method}
\usage{
confidenceBands(resultsSeeds, slot, ind)

\S4method{confidenceBands}{list,character,numeric}(resultsSeeds, slot, ind)

\S4method{confidenceBands}{list,character,missing}(resultsSeeds, slot, ind)

\S4method{confidenceBands}{resultsSeeds,character,missing}(resultsSeeds,
  slot, ind)
}
\arguments{
\item{resultsSeeds}{A object of the class 'resultsSeeds', which is returned from the algorithms.}

\item{slot}{Arguement of type character.}

\item{ind}{A numeric indicating the index of a 'resultsSeeds'-Object in a list. If not set the last listed object will be used.}
}
\description{
Get the estimated confidence bands for the bayesian method
}
