% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynElasticNet.R
\name{dynElasticNet}
\alias{dynElasticNet}
\title{estimating the optimal control using the dynamic elastic net}
\usage{
dynElasticNet(alphaStep, armijoBeta, x0, parameters, times, alpha1, alpha2,
  measData, constStr, STD, modelFunc, measFunc, modelInput, optW, origAUC,
  maxIteration, plotEsti, conjGrad)
}
\arguments{
\item{alphaStep}{starting value of the stepsize for the gradient descent, will be calculate to minimize the costfunction by backtracking algorithm}

\item{armijoBeta}{scaling of the alphaStep to find a approximatly optimal value for the stepsize}

\item{x0}{initial state of the ode system}

\item{parameters}{parameters of the ODE-system}

\item{times}{measurement times at which the estimations will be evaluated}

\item{alpha1}{L1 cost term skalar}

\item{alpha2}{L2 cost term skalar}

\item{measData}{measured values of the experiment}

\item{constStr}{a string that represents constrains, can be used to calculate a hidden input for a komponent that gradient is zero}

\item{STD}{standard deviation of the experiment; leave empty if unknown}

\item{modelFunc}{function that describes the ODE-system of the model}

\item{measFunc}{function that maps the states to the outputs}

\item{modelInput}{an dataset that discribes the external input of the system}

\item{optW}{vector that indicated at which knots of the network the alogrithm should estimate the hidden inputs}

\item{origAUC}{AUCs of the first optimisation; only used by the algorithm}

\item{maxIteration}{a upper bound for the maximal number of iterations}

\item{plotEsti}{boolean that controlls of the current estimates should be plotted}

\item{conjGrad}{boolean that indicates the usage of conjugate gradient method over the normal steepest descent}
}
\value{
A list containing the estimated hidden inputs, the AUCs, the estimated states and resulting measurements and the costfunction
}
\description{
estimating the optimal control using the dynamic elastic net
}
