% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsSeeds.R
\name{estiStates}
\alias{estiStates}
\alias{estiStates,list,numeric-method}
\alias{estiStates,list,missing-method}
\alias{estiStates,resultsSeeds,missing-method}
\title{Get the estimated states}
\usage{
estiStates(resultsSeeds, ind)

\S4method{estiStates}{list,numeric}(resultsSeeds, ind)

\S4method{estiStates}{list,missing}(resultsSeeds, ind)

\S4method{estiStates}{resultsSeeds,missing}(resultsSeeds, ind)
}
\arguments{
\item{resultsSeeds}{A object of the class resultsSeeds, which is returned from the algorithms.}

\item{ind}{A numeric indicating the index of a resultsSeeds-Object in a list. If not set the last listed object will be used.}
}
\value{
Dataframe containing the estimated states
}
\description{
Get the estimated states
}
\examples{

data(uvb_res)

estiStates(res)

}
