% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedyApproach.R
\name{greedyApproach}
\alias{greedyApproach}
\title{Greedy Approach Algorithm}
\usage{
greedyApproach(alphaStep, Beta, alpha1, alpha2, x0, optW, times, measFunc,
  measData, std, parameters, systemInput, modelFunc, greedyLogical,
  plotEstimates, conjGrad, cString)
}
\arguments{
\item{alphaStep}{the starting stepsize for the gradient descent
a fitting stepsize will be calculated based on a backtracking line search
if the algorithm converges to slow use a bigger stepsize}

\item{Beta}{skaling parameter for the backtracking to approximate the stepsize of the gradient descent. Is set to  0.8
if no value is given to the function}

\item{alpha1}{L1-norm parameter of the dynamic elastic net approach, is set to zero for this algorithm}

\item{alpha2}{L2-norm parameter of the dynamic elastic net approach
used for regulation purposes
set to NULL for a approximation of alpha2 - will results in a longer runtime}

\item{x0}{inital state of the ODE system}

\item{optW}{a vector that indicates for which knots of the network a input should be calculated}

\item{times}{time sequence for which output is wanted; the first value of times must be the initial time}

\item{measFunc}{a R-Function that is used for measurement of the states if the system is not completly
measurable; an empty argument will result in the assumption that the complete system is
measurable}

\item{measData}{a table that containts the measurements of the experiment; used to calculate the needed inputs}

\item{std}{standard deviation of the measurement; used to weight the errors of the estimates in the cost function}

\item{parameters}{vector or named vector that contains the parameters of the ODE equation}

\item{systemInput}{an dataset that discribes the external input of the system}

\item{modelFunc}{a R-Function that states the ODE system for which the hidden inputs should be calculated}

\item{greedyLogical}{a boolean that states if the greedy approach should be used; if set to FALSE the algorithm
will only use perform a calculation of the inputs for all knots without a sparse solution}

\item{plotEstimates}{boolean that indicated if the current estimate should be shown}

\item{conjGrad}{boolean that indicates the usage of conjugate gradient method over the normal steepest descent}

\item{cString}{a string that represents constrains, can be used to calculate a hidden input for a komponent that gradient is zero}
}
\value{
returns a results-object with default plot function. The plot shows the estimated best sparse fit
}
\description{
calculates controls based on a first optimisation with gradient descent; should result in a sparse vector
of hidden inputs
}
