% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsSeeds.R
\name{hiddenInputs}
\alias{hiddenInputs}
\title{Get the estimated hidden inputs}
\usage{
hiddenInputs(resultsSeeds, ind)
}
\arguments{
\item{resultsSeeds}{A object of the class 'resultsSeeds', which is returned from the algorithms.}

\item{ind}{A numeric indicating the index of a 'resultsSeeds'-Object in a list. If not set the last listed object will be used.}
}
\description{
Get the estimated hidden inputs
}
