% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSBML.R
\name{importSBML}
\alias{importSBML}
\title{Import SBML Models using the Bioconductor package 'rsbml'}
\usage{
importSBML(filename, times, y)
}
\arguments{
\item{filename}{name of the import file. Should be located in the working directory.}

\item{times}{timestep at which the function should be evaluated}

\item{y}{measurements of the model}
}
\value{
returns a odeModel-Object
}
\description{
A simple function for importing sbml models from a extensive markup language file.
}
