% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelClass.R
\docType{methods}
\name{nominalSol}
\alias{nominalSol}
\alias{nominalSol,odeModel,missing-method}
\alias{nominalSol,odeModel,numeric-method}
\title{Calculate the nominal solution of the model}
\usage{
nominalSol(odeModel, logTrans)

\S4method{nominalSol}{odeModel,missing}(odeModel, logTrans)

\S4method{nominalSol}{odeModel,numeric}(odeModel, logTrans)
}
\arguments{
\item{odeModel}{a object of the class ode model describing the experiment}

\item{logTrans}{a vector indicating which of the state vector componenets should be log transformed}
}
\description{
Calculate the nominal solution of the model
}
