% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelClass.R
\docType{methods}
\name{nominalSol}
\alias{nominalSol}
\alias{nominalSol,odeModel-method}
\title{Calculate the nominal solution of the model}
\usage{
nominalSol(odeModel)

\S4method{nominalSol}{odeModel}(odeModel)
}
\arguments{
\item{odeModel}{a object of the class ode model describing the experiment}
}
\description{
Calculate the nominal solution of the model
}
