% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelClass.R
\name{nominalSol}
\alias{nominalSol}
\alias{nominalSol,odeModel-method}
\title{Calculate the nominal solution of the model}
\usage{
nominalSol(odeModel)

\S4method{nominalSol}{odeModel}(odeModel)
}
\arguments{
\item{odeModel}{a object of the class ode model describing the experiment}
}
\value{
a matrix with the numeric solution to the nominal ode equation
}
\description{
After an model is defined it can be evaluated. This returns the numerical solution
for the state equation before hidden inputs are calculated.
}
\examples{

lotka_voltera <- function (Time, State, Pars) {
with(as.list(c(State, Pars)), {
 dx1 = x1*(alpha - beta*x2)
  dx2 = -x2*(gamma - delta*x1)
 return(list(c(dx, dy)))
})
}

pars <- c(alpha = 2, beta = .5, gamma = .2, delta = .6)
init_state <- c(x = 10, y = 10)
time <- seq(0, 100, by = 1)
lotVolModel = odeModel(func = lotka_voltera, parms = pars, times = time, y = init_state)
nominalSol(lotVolModel) 

}
