% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelClass.R
\name{setModelEquation}
\alias{setModelEquation}
\title{Set the model equation}
\usage{
setModelEquation(theObject, func)
}
\arguments{
\item{theObject}{an object of the class modelClass}

\item{func}{function describing the ode equation of the model}
}
\description{
Set the model equation of the system. Has to be a function that can be used with the deSolve package
}
