% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsSeeds.R
\name{plotAnno}
\alias{plotAnno}
\alias{plotAnno,resultsSeeds-method}
\alias{plotAnno,list-method}
\title{Create annotated plot}
\usage{
plotAnno(x, stateAnno, measAnno)

\S4method{plotAnno}{resultsSeeds}(x, stateAnno, measAnno)

\S4method{plotAnno}{list}(x, stateAnno, measAnno)
}
\arguments{
\item{x}{an object of type resultsSeeds which contains the results of the algorithms}

\item{stateAnno}{a character vector describing the names of the states}

\item{measAnno}{a character vector describing the names of the measurements}
}
\value{
Plots of the results with the provided annotation
}
\description{
Create a annotated plot with given state and measurement names. The plots are
equal to the output of the normal plot function.
}
\examples{

data(uvb_res)

statesAnno <- c("x1", "x2", "x3", "x4", "x5", "x6", "x7", "x8", "x9", "x10", "x11", "x12", "x13")
measurAnno <- c("y1", "y2", "y3", "y4", "y5")

plotAnno(res[[2]], stateAnno = statesAnno, measAnno =  measurAnno)

}
