% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsSeeds.R
\name{print,resultsSeeds-method}
\alias{print,resultsSeeds-method}
\alias{print,resultsSeeds}
\title{A default printing function for the resultsSeeds class}
\usage{
\S4method{print}{resultsSeeds}(x)
}
\arguments{
\item{x}{an object of the class resultsSeeds}
}
\value{
Returns a short summary of the important results
}
\description{
This function overwrites the default print function and is used for objects 
of the class resultsSeeds. The print function gives the basic information about
the results seeds object. The default printout is the estimated states and 
the calculated hidden inputs
}
\examples{
data(ubv_res)

plot(res[[2]])


}
