% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds.R
\docType{package}
\name{seeds-package}
\alias{seeds}
\alias{seeds-package}
\title{seeds: Estimate hidden inputs using the method of the dynamic elastic net}
\description{
A package combining two different algorithms for calculating hidden
inputs. These hidden inputs can be interpreted as a controll that tries to minimize the
discrepancies between a given model and a given measurements. The method is
also called the 'dynamic elastic net'. The related work can be found in doi:10.1038/srep20772
}
\details{
Details

The conjugate gradient based algorithm uses a greedy algorithm to estimate a 
sparse controll that tries to minimize the discrepancies between a given 
'nominal model given the measurements (e.g from an experiment). The algorithm
uses \pkg{deSolve} to calculate the hidden inputs w based on the adjoint equations
of the ODE-System. 

The adjoint equations are calculated using \pkg{\code{\link{Deriv}}}. For the usage of the 
algorithm please look into the examples and documentation given for the
functions 
\describe{
   \item{\code{\link{greedyApproach}}}{a greedy algorithm to calculate a sparse control}
   \item{\code{\link{dynElasticNet}}}{a conjugate gradient optimal control algorithm with L2 regulation} 
}
}
\author{
\strong{Maintainer}: Tobias Newmiwaka \email{tobias.newmiwaka@gmail.com}

}
