% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelClass.R
\docType{methods}
\name{setInput}
\alias{setInput}
\alias{setInput,odeModel-method}
\title{Set the inputs of the model.}
\usage{
setInput(theObject, input)

\S4method{setInput}{odeModel}(theObject, input)
}
\arguments{
\item{theObject}{an object of the class modelClass}

\item{input}{function describing the ode equation of the model}
}
\description{
Set the inputs of the model.
}
