% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelClass.R
\name{setMeas}
\alias{setMeas}
\alias{setMeas,odeModel-method}
\title{set measurements of the model}
\usage{
setMeas(odeModel, meas)

\S4method{setMeas}{odeModel}(odeModel, meas)
}
\arguments{
\item{odeModel}{an object of the class odeModel}

\item{meas}{measurements of the model, a matrix with measurements of the model
and the corresponding time values}
}
\value{
an object of odeModel
}
\description{
The odeModel object stores all important information. Measurements of the objects can be set
directly by adressing the slot, or with this function.
}
\examples{

data(uvbData)
data(uvbModel)

measurements <- uvbData[,1:6]

newModel <- setMeas(odeModel = uvbModel, meas = measurements)

}
