% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelClass.R
\name{setMeasFunc}
\alias{setMeasFunc}
\title{Set the measurement equation for the model}
\usage{
setMeasFunc(theObject, measFunc, costum)
}
\arguments{
\item{theObject}{an object of the class odeModel}

\item{measFunc}{measurement function of the model. Has to be a R functions.}

\item{costum}{costum indexing for the measurement function}
}
\description{
Set the measurement equation for the model
}
