% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelClass.R
\name{setParms}
\alias{setParms}
\title{set the model parameters}
\usage{
setParms(theObject, parms)
}
\arguments{
\item{theObject}{an object of the class odeModel}

\item{parms}{a vector containing the parmeters of the model}
}
\description{
set the model parameters
}
