% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelClass.R
\name{setY}
\alias{setY}
\title{Set the vector with the initial (state) values}
\usage{
setY(theObject, y)
}
\arguments{
\item{theObject}{an object of the class odeModel}

\item{y}{vector with the initial values}
}
\description{
Set the vector with the initial (state) values
}
